refactor: restructure domain [deps-change]
- Moved CLI options and configuration schemas from `domain/cli-options.js` to `domain/models/options.js`.
- Introduced new error classes in `domain/models/errors.ts` for better error handling.
- Created a new `domain/constants/package-info.ts` to manage package metadata.
- Removed deprecated files: `application/markdown-builder.ts`, `application/mermaid-file-processor.ts`, and `application/process-result-types.ts`.
- Implemented a new `application/use-cases/convert-files.ts` that orchestrates the file conversion process with improved error handling.
- Added a `batch-processor.service.ts` for handling batch file processing and result aggregation.
- Created a `file-storage.service.ts` to abstract file operations and improve maintainability.
- Updated CLI interaction in `presentation/cli.ts` to directly handle conversion results and errors.
- Removed the old CLI adapter and console reporter, integrating their functionalities into the main CLI command.
- Added tests for the options merger to ensure correct behavior when merging CLI and config options.
