+++ src/domain/models/errors.ts
