+++ src/infrastructure/adapters/file-system.adapter.ts
@@ -1,0 +1,39 @@
+/**
+ * File system adapter
+ *
+ * This adapter provides a thin wrapper around Node.js file system operations.
+ * It abstracts the technical details of file I/O operations.
+ */
+
+import { mkdir, readFile, unlink, writeFile } from "node:fs/promises";
+
+/**
+ * Read file content as UTF-8
+ */
+export async function readTextFile(filePath: string): Promise<string> {
+  return await readFile(filePath, "utf-8");
+}
+
+/**
+ * Write text content to file as UTF-8
+ */
+export async function writeTextFile(
+  filePath: string,
+  content: string,
+): Promise<void> {
+  await writeFile(filePath, content, "utf-8");
+}
+
+/**
+ * Delete a file from the file system
+ */
+export async function deleteFile(filePath: string): Promise<void> {
+  await unlink(filePath);
+}
+
+/**
+ * Create a directory (including parent directories)
+ */
+export async function createDirectory(path: string): Promise<void> {
+  await mkdir(path, { recursive: true });
+}
